pret a etre valide

+++ ex02/Array.hpp
@@ -8,21 +8,21 @@ template <typename T>
 class	Array
 {
 	private:
-		size_t	_size;
-		T	*_arr;
+		unsigned int	_size;
+		T		*_arr;
 
 	public:
-			Array(void);
-			~Array(void);
+				Array(void);
+				~Array(void);
 
-			Array(unsigned int n);
-			Array(const Array& other);
-		Array&	operator=(const Array& other);
+				Array(unsigned int n);
+				Array(const Array& other);
+		Array&		operator=(const Array& other);
 
-		T&	operator[](unsigned int index);
-		size_t	size(void) const;
+		T&		operator[](unsigned int index);
+		unsigned int	size(void) const;
 
-		class	OutOfRange : public std::exception
+		class		OutOfRange : public std::exception
 		{
 			const char	*what(void) const throw();
 		};
